Add: Certificate info for TippingPoint alerts
The details of alerts using the TippingPoint SMS method now shows
information about the TLS certificate.

This way it is possible to check if the correct certificate is
configured or not.
