+++ src/web/components/certinfo/certinfo.js
@@ -1,0 +1,85 @@
+/* Copyright (C) 2023 Greenbone AG
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Affero General Public License
+ * as published by the Free Software Foundation, either version 3
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from 'react';
+
+import _ from 'gmp/locale';
+
+import {isDefined} from 'gmp/utils/identity';
+
+import PropTypes from 'web/utils/proptypes';
+
+import DateTime from 'web/components/date/datetime';
+
+import {Col} from 'web/entity/page';
+
+import InfoTable from 'web/components/table/infotable';
+import TableBody from 'web/components/table/body';
+import TableData from 'web/components/table/data';
+import TableRow from 'web/components/table/row';
+
+const CertInfo = ({info}) => {
+  const {activationTime, expirationTime, issuer, md5_fingerprint} = info;
+  return (
+    <InfoTable>
+      <colgroup>
+        <Col width="10%" />
+        <Col width="90%" />
+      </colgroup>
+      <TableBody>
+        <TableRow>
+          <TableData>{_('Activation')}</TableData>
+          <TableData>
+            {isDefined(activationTime) ? (
+              <DateTime date={activationTime} />
+            ) : (
+              _('N/A')
+            )}
+          </TableData>
+        </TableRow>
+
+        <TableRow>
+          <TableData>{_('Expiration')}</TableData>
+          <TableData>
+            {isDefined(expirationTime) ? (
+              <DateTime date={expirationTime} />
+            ) : (
+              _('N/A')
+            )}
+          </TableData>
+        </TableRow>
+
+        <TableRow>
+          <TableData>{_('MD5 Fingerprint')}</TableData>
+          <TableData>{md5_fingerprint}</TableData>
+        </TableRow>
+
+        <TableRow>
+          <TableData>{_('Issuer')}</TableData>
+          <TableData>{issuer}</TableData>
+        </TableRow>
+      </TableBody>
+    </InfoTable>
+  );
+};
+
+CertInfo.propTypes = {
+  info: PropTypes.object.isRequired,
+};
+
+export default CertInfo;
