+++ src/web/pages/scanners/details.js
@@ -28,66 +28,18 @@ import PropTypes from 'web/utils/proptypes';
 
 import DetailsBlock from 'web/entity/block';
 
+import CertInfo from 'web/components/certinfo/certinfo';
+
 import Layout from 'web/components/layout/layout';
 
 import DetailsLink from 'web/components/link/detailslink';
 
-import DateTime from 'web/components/date/datetime';
 import InfoTable from 'web/components/table/infotable';
 import TableBody from 'web/components/table/body';
 import TableData, {TableDataAlignTop} from 'web/components/table/data';
 import TableRow from 'web/components/table/row';
 
 import {Col} from 'web/entity/page';
-
-const CertInfo = ({info}) => {
-  const {activationTime, expirationTime, issuer, md5_fingerprint} = info;
-  return (
-    <InfoTable>
-      <colgroup>
-        <Col width="10%" />
-        <Col width="90%" />
-      </colgroup>
-      <TableBody>
-        <TableRow>
-          <TableData>{_('Activation')}</TableData>
-          <TableData>
-            {isDefined(activationTime) ? (
-              <DateTime date={activationTime} />
-            ) : (
-              _('N/A')
-            )}
-          </TableData>
-        </TableRow>
-
-        <TableRow>
-          <TableData>{_('Expiration')}</TableData>
-          <TableData>
-            {isDefined(expirationTime) ? (
-              <DateTime date={expirationTime} />
-            ) : (
-              _('N/A')
-            )}
-          </TableData>
-        </TableRow>
-
-        <TableRow>
-          <TableData>{_('MD5 Fingerprint')}</TableData>
-          <TableData>{md5_fingerprint}</TableData>
-        </TableRow>
-
-        <TableRow>
-          <TableData>{_('Issuer')}</TableData>
-          <TableData>{issuer}</TableData>
-        </TableRow>
-      </TableBody>
-    </InfoTable>
-  );
-};
-
-CertInfo.propTypes = {
-  info: PropTypes.object.isRequired,
-};
 
 const ScannerDetails = ({entity}) => {
   const {
